requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0
alembic>=1.11
pydantic>=2.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
